# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apatris Compliance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Attachment {
  id: string;
  url: string;
  filename: string;
  /** @nullable */
  size?: number | null;
  /** @nullable */
  type?: string | null;
}

export interface Worker {
  id: string;
  name: string;
  /** TIG, MIG, or ARC */
  specialization: string;
  /**
   * ISO date string
   * @nullable
   */
  trcExpiry?: string | null;
  /**
   * ISO date string
   * @nullable
   */
  workPermitExpiry?: string | null;
  /**
   * Active, Expired, Pending
   * @nullable
   */
  bhpStatus?: string | null;
  /**
   * ISO date string
   * @nullable
   */
  contractEndDate?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** critical, warning, compliant, non-compliant */
  complianceStatus: string;
  /** @nullable */
  daysUntilNextExpiry?: number | null;
  passportAttachments: Attachment[];
  contractAttachments: Attachment[];
}

export interface WorkerListResponse {
  workers: Worker[];
  total: number;
}

export interface UpdateWorkerRequest {
  /** @nullable */
  trcExpiry?: string | null;
  /** @nullable */
  workPermitExpiry?: string | null;
  /** @nullable */
  bhpStatus?: string | null;
  /** @nullable */
  contractEndDate?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
}

export interface NotifyWorkerRequest {
  message: string;
  /** email or sms */
  channel?: string;
}

export interface NotifyWorkerResponse {
  success: boolean;
  message: string;
}

export interface WorkerStats {
  total: number;
  critical: number;
  warning: number;
  compliant: number;
  nonCompliant: number;
}

export interface ExpiringDocument {
  workerId: string;
  workerName: string;
  specialization: string;
  documentType: string;
  expiryDate: string;
  daysUntilExpiry: number;
  status: string;
}

export interface ComplianceReport {
  generatedAt: string;
  totalWorkers: number;
  expiringThisWeek: ExpiringDocument[];
  critical: ExpiringDocument[];
  warning: ExpiringDocument[];
  summary: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetWorkersParams = {
  /**
   * Search by name
   */
  search?: string;
  /**
   * Filter by specialization (TIG, MIG, ARC)
   */
  specialization?: string;
  /**
   * Filter by compliance status (critical, warning, compliant, non-compliant)
   */
  status?: string;
};