# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Apatris Compliance Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: workers
    description: Worker management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /workers:
    get:
      operationId: getWorkers
      tags: [workers]
      summary: Get all workers
      description: Fetches all welder records from Airtable
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by name
        - name: specialization
          in: query
          required: false
          schema:
            type: string
          description: Filter by specialization (TIG, MIG, ARC)
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by compliance status (critical, warning, compliant, non-compliant)
      responses:
        "200":
          description: List of workers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerListResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      summary: Get a worker by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Worker details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateWorker
      tags: [workers]
      summary: Update a worker record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkerRequest"
      responses:
        "200":
          description: Updated worker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /workers/{id}/notify:
    post:
      operationId: notifyWorker
      tags: [workers]
      summary: Send notification to worker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotifyWorkerRequest"
      responses:
        "200":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotifyWorkerResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /workers/report:
    get:
      operationId: getComplianceReport
      tags: [workers]
      summary: Generate compliance report
      description: Summarizes all workers with expiring documents
      responses:
        "200":
          description: Compliance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceReport"

  /workers/stats:
    get:
      operationId: getWorkerStats
      tags: [workers]
      summary: Get worker compliance statistics
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Attachment:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        filename:
          type: string
        size:
          type: ["integer", "null"]
        type:
          type: ["string", "null"]
      required:
        - id
        - url
        - filename

    Worker:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        specialization:
          type: string
          description: "TIG, MIG, or ARC"
        trcExpiry:
          type: ["string", "null"]
          description: "ISO date string"
        workPermitExpiry:
          type: ["string", "null"]
          description: "ISO date string"
        bhpStatus:
          type: ["string", "null"]
          description: "Active, Expired, Pending"
        contractEndDate:
          type: ["string", "null"]
          description: "ISO date string"
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        complianceStatus:
          type: string
          description: "critical, warning, compliant, non-compliant"
        daysUntilNextExpiry:
          type: ["integer", "null"]
        passportAttachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        contractAttachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
      required:
        - id
        - name
        - specialization
        - complianceStatus
        - passportAttachments
        - contractAttachments

    WorkerListResponse:
      type: object
      properties:
        workers:
          type: array
          items:
            $ref: "#/components/schemas/Worker"
        total:
          type: integer
      required:
        - workers
        - total

    UpdateWorkerRequest:
      type: object
      properties:
        trcExpiry:
          type: ["string", "null"]
        workPermitExpiry:
          type: ["string", "null"]
        bhpStatus:
          type: ["string", "null"]
        contractEndDate:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]

    NotifyWorkerRequest:
      type: object
      properties:
        message:
          type: string
        channel:
          type: string
          description: "email or sms"
      required:
        - message

    NotifyWorkerResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    WorkerStats:
      type: object
      properties:
        total:
          type: integer
        critical:
          type: integer
        warning:
          type: integer
        compliant:
          type: integer
        nonCompliant:
          type: integer
      required:
        - total
        - critical
        - warning
        - compliant
        - nonCompliant

    ExpiringDocument:
      type: object
      properties:
        workerId:
          type: string
        workerName:
          type: string
        specialization:
          type: string
        documentType:
          type: string
        expiryDate:
          type: string
        daysUntilExpiry:
          type: integer
        status:
          type: string
      required:
        - workerId
        - workerName
        - specialization
        - documentType
        - expiryDate
        - daysUntilExpiry
        - status

    ComplianceReport:
      type: object
      properties:
        generatedAt:
          type: string
        totalWorkers:
          type: integer
        expiringThisWeek:
          type: array
          items:
            $ref: "#/components/schemas/ExpiringDocument"
        critical:
          type: array
          items:
            $ref: "#/components/schemas/ExpiringDocument"
        warning:
          type: array
          items:
            $ref: "#/components/schemas/ExpiringDocument"
        summary:
          type: string
      required:
        - generatedAt
        - totalWorkers
        - expiringThisWeek
        - critical
        - warning
        - summary

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error