# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apatris Compliance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Fetches all welder records from Airtable
 * @summary Get all workers
 */
export const GetWorkersQueryParams = zod.object({
  search: zod.coerce.string().optional().describe("Search by name"),
  specialization: zod.coerce
    .string()
    .optional()
    .describe("Filter by specialization (TIG, MIG, ARC)"),
  status: zod.coerce
    .string()
    .optional()
    .describe(
      "Filter by compliance status (critical, warning, compliant, non-compliant)",
    ),
});

export const GetWorkersResponse = zod.object({
  workers: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      specialization: zod.string().describe("TIG, MIG, or ARC"),
      trcExpiry: zod.string().nullish().describe("ISO date string"),
      workPermitExpiry: zod.string().nullish().describe("ISO date string"),
      bhpStatus: zod.string().nullish().describe("Active, Expired, Pending"),
      contractEndDate: zod.string().nullish().describe("ISO date string"),
      email: zod.string().nullish(),
      phone: zod.string().nullish(),
      complianceStatus: zod
        .string()
        .describe("critical, warning, compliant, non-compliant"),
      daysUntilNextExpiry: zod.number().nullish(),
      passportAttachments: zod.array(
        zod.object({
          id: zod.string(),
          url: zod.string(),
          filename: zod.string(),
          size: zod.number().nullish(),
          type: zod.string().nullish(),
        }),
      ),
      contractAttachments: zod.array(
        zod.object({
          id: zod.string(),
          url: zod.string(),
          filename: zod.string(),
          size: zod.number().nullish(),
          type: zod.string().nullish(),
        }),
      ),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get a worker by ID
 */
export const GetWorkerParams = zod.object({
  id: zod.coerce.string(),
});

export const GetWorkerResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  specialization: zod.string().describe("TIG, MIG, or ARC"),
  trcExpiry: zod.string().nullish().describe("ISO date string"),
  workPermitExpiry: zod.string().nullish().describe("ISO date string"),
  bhpStatus: zod.string().nullish().describe("Active, Expired, Pending"),
  contractEndDate: zod.string().nullish().describe("ISO date string"),
  email: zod.string().nullish(),
  phone: zod.string().nullish(),
  complianceStatus: zod
    .string()
    .describe("critical, warning, compliant, non-compliant"),
  daysUntilNextExpiry: zod.number().nullish(),
  passportAttachments: zod.array(
    zod.object({
      id: zod.string(),
      url: zod.string(),
      filename: zod.string(),
      size: zod.number().nullish(),
      type: zod.string().nullish(),
    }),
  ),
  contractAttachments: zod.array(
    zod.object({
      id: zod.string(),
      url: zod.string(),
      filename: zod.string(),
      size: zod.number().nullish(),
      type: zod.string().nullish(),
    }),
  ),
});

/**
 * @summary Update a worker record
 */
export const UpdateWorkerParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateWorkerBody = zod.object({
  trcExpiry: zod.string().nullish(),
  workPermitExpiry: zod.string().nullish(),
  bhpStatus: zod.string().nullish(),
  contractEndDate: zod.string().nullish(),
  email: zod.string().nullish(),
  phone: zod.string().nullish(),
});

export const UpdateWorkerResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  specialization: zod.string().describe("TIG, MIG, or ARC"),
  trcExpiry: zod.string().nullish().describe("ISO date string"),
  workPermitExpiry: zod.string().nullish().describe("ISO date string"),
  bhpStatus: zod.string().nullish().describe("Active, Expired, Pending"),
  contractEndDate: zod.string().nullish().describe("ISO date string"),
  email: zod.string().nullish(),
  phone: zod.string().nullish(),
  complianceStatus: zod
    .string()
    .describe("critical, warning, compliant, non-compliant"),
  daysUntilNextExpiry: zod.number().nullish(),
  passportAttachments: zod.array(
    zod.object({
      id: zod.string(),
      url: zod.string(),
      filename: zod.string(),
      size: zod.number().nullish(),
      type: zod.string().nullish(),
    }),
  ),
  contractAttachments: zod.array(
    zod.object({
      id: zod.string(),
      url: zod.string(),
      filename: zod.string(),
      size: zod.number().nullish(),
      type: zod.string().nullish(),
    }),
  ),
});

/**
 * @summary Send notification to worker
 */
export const NotifyWorkerParams = zod.object({
  id: zod.coerce.string(),
});

export const NotifyWorkerBody = zod.object({
  message: zod.string(),
  channel: zod.string().optional().describe("email or sms"),
});

export const NotifyWorkerResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * Summarizes all workers with expiring documents
 * @summary Generate compliance report
 */
export const GetComplianceReportResponse = zod.object({
  generatedAt: zod.string(),
  totalWorkers: zod.number(),
  expiringThisWeek: zod.array(
    zod.object({
      workerId: zod.string(),
      workerName: zod.string(),
      specialization: zod.string(),
      documentType: zod.string(),
      expiryDate: zod.string(),
      daysUntilExpiry: zod.number(),
      status: zod.string(),
    }),
  ),
  critical: zod.array(
    zod.object({
      workerId: zod.string(),
      workerName: zod.string(),
      specialization: zod.string(),
      documentType: zod.string(),
      expiryDate: zod.string(),
      daysUntilExpiry: zod.number(),
      status: zod.string(),
    }),
  ),
  warning: zod.array(
    zod.object({
      workerId: zod.string(),
      workerName: zod.string(),
      specialization: zod.string(),
      documentType: zod.string(),
      expiryDate: zod.string(),
      daysUntilExpiry: zod.number(),
      status: zod.string(),
    }),
  ),
  summary: zod.string(),
});

/**
 * @summary Get worker compliance statistics
 */
export const GetWorkerStatsResponse = zod.object({
  total: zod.number(),
  critical: zod.number(),
  warning: zod.number(),
  compliant: zod.number(),
  nonCompliant: zod.number(),
});