# lib/api-zod/src/generated/types/attachment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apatris Compliance Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Attachment {
  id: string;
  url: string;
  filename: string;
  /** @nullable */
  size?: number | null;
  /** @nullable */
  type?: string | null;
}