# lib/api-zod/src/generated/types/complianceReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apatris Compliance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpiringDocument } from "./expiringDocument";

export interface ComplianceReport {
  generatedAt: string;
  totalWorkers: number;
  expiringThisWeek: ExpiringDocument[];
  critical: ExpiringDocument[];
  warning: ExpiringDocument[];
  summary: string;
}