# lib/api-zod/src/generated/types/expiringDocument.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apatris Compliance Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpiringDocument {
  workerId: string;
  workerName: string;
  specialization: string;
  documentType: string;
  expiryDate: string;
  daysUntilExpiry: number;
  status: string;
}