# lib/api-zod/src/generated/types/notifyWorkerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apatris Compliance Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface NotifyWorkerRequest {
  message: string;
  /** email or sms */
  channel?: string;
}