# lib/api-zod/src/generated/types/updateWorkerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apatris Compliance Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateWorkerRequest {
  /** @nullable */
  trcExpiry?: string | null;
  /** @nullable */
  workPermitExpiry?: string | null;
  /** @nullable */
  bhpStatus?: string | null;
  /** @nullable */
  contractEndDate?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
}