# lib/api-zod/src/generated/types/worker.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apatris Compliance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Attachment } from "./attachment";

export interface Worker {
  id: string;
  name: string;
  /** TIG, MIG, or ARC */
  specialization: string;
  /**
   * ISO date string
   * @nullable
   */
  trcExpiry?: string | null;
  /**
   * ISO date string
   * @nullable
   */
  workPermitExpiry?: string | null;
  /**
   * Active, Expired, Pending
   * @nullable
   */
  bhpStatus?: string | null;
  /**
   * ISO date string
   * @nullable
   */
  contractEndDate?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** critical, warning, compliant, non-compliant */
  complianceStatus: string;
  /** @nullable */
  daysUntilNextExpiry?: number | null;
  passportAttachments: Attachment[];
  contractAttachments: Attachment[];
}