# lib/api-zod/src/generated/types/workerListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apatris Compliance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Worker } from "./worker";

export interface WorkerListResponse {
  workers: Worker[];
  total: number;
}